Fix the inverted `Eye extends SparseEye` via a shared `EyeBase`.
Dense `Eye` (tf.eye) extended sparse `SparseEye` (tf.sparse.eye) and
borrowed its num_rows x num_columns shape logic via `super.getShapes()`
— an inverted, code-reuse-not-is-a relationship (wala/ML#514). Neither
is a kind of the other: `tf.eye` takes a `batch_shape` that
`tf.sparse.eye` lacks. Extract the shared shape construction into an
abstract `EyeBase extends TensorTypeAllocator`; `Eye` and `SparseEye`
both extend it, each supplying its own `dtype` parameter position, and
`Eye` adds the `batch_shape` prepend. This also lifts `SparseEye` off
the `extends Ones` chain.

Behavior-preserving: the shape logic is relocated verbatim. Full suite
green (969 tests). With dense `Eye` no longer extending the sparse
generator, the `Eye` sparse-marker opt-out added in wala/ML#588 becomes
unnecessary once that work rebases onto this.
