+++ com.ibm.wala.cast.python.ml/source/com/ibm/wala/cast/python/ml/client/SparseEye.java
@@ -1,24 +1,18 @@
 package com.ibm.wala.cast.python.ml.client;
 
-import com.ibm.wala.cast.python.ml.types.TensorType.Dimension;
-import com.ibm.wala.cast.python.ml.types.TensorType.NumericDim;
 import com.ibm.wala.ipa.callgraph.CGNode;
-import com.ibm.wala.ipa.callgraph.propagation.InstanceKey;
 import com.ibm.wala.ipa.callgraph.propagation.PointsToSetVariable;
-import com.ibm.wala.ipa.callgraph.propagation.PropagationCallGraphBuilder;
-import com.ibm.wala.util.collections.HashSetFactory;
-import com.ibm.wala.util.intset.OrdinalSet;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Locale;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 
-public class SparseEye extends Ones {
-
-  private static final int SHAPE_PARAMETER_POSITION = UNDEFINED_PARAMETER_POSITION;
+/**
+ * A generator for the sparse identity matrix produced by {@code tf.sparse.eye}. The {@code
+ * num_rows} &times; {@code num_columns} shape construction lives in {@link EyeBase}; this subclass
+ * supplies the sparse signature's {@code dtype} parameter position (no {@code batch_shape}, unlike
+ * {@link Eye}).
+ *
+ * @see <a href="https://www.tensorflow.org/api_docs/python/tf/sparse/eye">tf.sparse.eye API</a>.
+ */
+public class SparseEye extends EyeBase {
 
   protected enum Parameters {
     NUM_ROWS,
@@ -43,134 +37,6 @@ public class SparseEye extends Ones {
     super(node);
   }
 
-  /**
-   * Retrieves the possible integer values for a specific argument.
-   *
-   * <p>It first checks the symbol table for constant values associated with the argument's value
-   * number. If not found, it falls back to the points-to analysis results to find potential integer
-   * values.
-   *
-   * @param builder the propagation call graph builder
-   * @param paramPosition the positional index of the argument
-   * @param paramName the keyword name of the argument
-   * @return a set of optional integers representing the possible values. An empty optional
-   *     indicates a null or non-integer value.
-   */
-  protected Set<Optional<Integer>> getPossibleArgumentValues(
-      PropagationCallGraphBuilder builder, int paramPosition, String paramName) {
-    int valNum = this.getArgumentValueNumber(builder, paramPosition, paramName, true);
-    if (valNum <= 0) return HashSetFactory.make();
-
-    if (this.getNode().getIR() != null
-        && this.getNode().getIR().getSymbolTable().isConstant(valNum)) {
-      Object c = this.getNode().getIR().getSymbolTable().getConstantValue(valNum);
-      Set<Optional<Integer>> ret = HashSetFactory.make();
-      if (c instanceof Number) {
-        ret.add(Optional.of(((Number) c).intValue()));
-      } else if (c == null) {
-        ret.add(Optional.empty());
-      }
-      if (!ret.isEmpty()) {
-        return ret;
-      }
-    }
-
-    OrdinalSet<InstanceKey> pts = this.getArgumentPointsToSet(builder, paramPosition, paramName);
-
-    if (pts == null || pts.isEmpty())
-      // Fallback to default (empty).
-      return HashSetFactory.make();
-
-    return StreamSupport.stream(pts.spliterator(), false)
-        .map(SparseEye::getIntValueFromInstanceKey)
-        .collect(Collectors.toSet());
-  }
-
-  @Override
-  protected Set<List<Dimension<?>>> getShapes(PropagationCallGraphBuilder builder) {
-    Set<List<Dimension<?>>> ret = HashSetFactory.make();
-    Set<Optional<Integer>> numRows = this.getNumberOfRows(builder);
-    Set<Optional<Integer>> numColumns = this.getNumberOfColumns(builder);
-
-    for (Optional<Integer> nRow : numRows) {
-      if (numColumns.isEmpty())
-        // If numColumns is not provided, it defaults to numRows.
-        for (Optional<Integer> nCol : numRows)
-          // Build the shape using nRow and nCol.
-          numColumns.add(nCol);
-
-      for (Optional<Integer> nCol : numColumns)
-        if (nCol.isEmpty()) {
-          // If numColumns is not provided, it defaults to numRows.
-          for (Optional<Integer> nCol2 : numRows) {
-            // Build the shape using nRow and nCol.
-            List<Dimension<?>> shape = new ArrayList<>();
-
-            NumericDim rowDim = new NumericDim(nRow.get());
-            NumericDim colDim = new NumericDim(nCol2.get());
-
-            shape.add(rowDim);
-            shape.add(colDim);
-
-            ret.add(shape);
-          }
-        } else {
-          List<Dimension<?>> shape = new ArrayList<>();
-
-          NumericDim rowDim = new NumericDim(nRow.get());
-          NumericDim colDim = new NumericDim(nCol.get());
-
-          shape.add(rowDim);
-          shape.add(colDim);
-
-          ret.add(shape);
-        }
-    }
-
-    return ret;
-  }
-
-  private Set<Optional<Integer>> getNumberOfRows(PropagationCallGraphBuilder builder) {
-    Set<Optional<Integer>> values =
-        this.getPossibleArgumentValues(
-            builder, this.getNumRowsParameterPosition(), this.getNumRowsParameterName());
-
-    if (values == null || values.isEmpty())
-      throw new IllegalStateException("The num_rows parameter is required for tf.eye().");
-
-    return values;
-  }
-
-  private Set<Optional<Integer>> getNumberOfColumns(PropagationCallGraphBuilder builder) {
-    return this.getPossibleArgumentValues(
-        builder, this.getNumColumnsParameterPosition(), this.getNumColumnsParameterName());
-  }
-
-  @Override
-  protected int getShapeParameterPosition() {
-    return SHAPE_PARAMETER_POSITION;
-  }
-
-  protected int getNumRowsParameterPosition() {
-    return Parameters.NUM_ROWS.getIndex();
-  }
-
-  protected String getNumRowsParameterName() {
-    return Parameters.NUM_ROWS.getName();
-  }
-
-  protected int getNumRowsArgumentValueNumber() {
-    return this.getArgumentValueNumber(this.getNumRowsParameterPosition());
-  }
-
-  protected int getNumColumnsParameterPosition() {
-    return Parameters.NUM_COLUMNS.getIndex();
-  }
-
-  protected String getNumColumnsParameterName() {
-    return Parameters.NUM_COLUMNS.getName();
-  }
-
   @Override
   protected int getDTypeParameterPosition() {
     return Parameters.DTYPE.getIndex();
@@ -178,10 +44,5 @@ public class SparseEye extends Ones {
 
   protected String getDTypeParameterName() {
     return Parameters.DTYPE.getName();
-  }
-
-  protected int getNumColumnsArgumentValueNumber(PropagationCallGraphBuilder builder) {
-    return this.getArgumentValueNumber(
-        builder, this.getNumColumnsParameterPosition(), this.getNumColumnsParameterName(), true);
   }
 }
